fix: main.ts의 앱 설정과 서버 실행 로직 분리
